OSM->Thunderforest map, use local providers repo
Create a mapping from the OSM Plugin map types it is expecting, to the
ones that Thunderforest provides, as they are not equivalent.

One of them - satellite - has no mapping to thunderforest's provided map
tile sets - it provides no satellite tiles -- in this case I just mapped
it to the same name, which results in no tiles loaded when that is
selected.

Two others "terrain" and "hiking" only have a single map tile set
provided by Thunderforest that most closely matches -- the "outdoors"
map tile set - so both of these map to this same tile set.

Additionally, this change removes the alternate approach to loading
tiles without providing a local server, and removes the dynamic loading
of `Map` and `Plugin`, as it is not needed with the local server
approach.