Prevent QML Engine from creating second AppConfig
According to docs (and empirical evidence - I saw this myself), the QML engine
will construct a QML_SINGLETON instance if the type has a default constructor,
even if there is a static factory function of the signature
`T *create(QQmlEngine *, QJSEngine *)`.

I have a factory function with the right signature for AppConfig, however,
because I had a default constructor, the engine was constructing a second
instance instead of calling the `create` factory function.

This is solved by making the parent argument mandatory instead of optional.