Load thunderforest maps
This loads up thunderforest maps.

Presently it has two different ways it does it in the code, with the
simpler way enabled presently. The http server is not actually being
used at the moment, but it is started.

The problem with this approach right now is that I haven't been able to
switch between different tilesets.